requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.1.0
beautifulsoup4==4.11.1
cachetools==5.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
chinese-converter==1.0.2
Django==4.1.4
django-cors-headers==3.14.0
djangorestframework==3.14.0
frozenlist==1.3.3
future==0.18.2
idna==3.4
importlib-resources==5.10.1
Js2Py==0.74
lxml==4.9.2
lzstring==1.0.4
multidict==6.0.3
nest-asyncio==1.5.6
pyjsparser==2.7.1
python-dotenv==0.21.0
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.13
yarl==1.8.2
zipp==3.11.0
